python-backend/requirements.txt
# Core
Flask>=3.0.2
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
Flask-Limiter>=3.5.1
Flask-Cors>=4.0.1
gunicorn>=21.2.0

# Database
SQLAlchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Auth
PyJWT>=2.8.0
Werkzeug>=3.0.1
python-dotenv>=1.0.1

# Storage
cloudinary>=1.39.1
boto3>=1.34.21

# AI (Only Gemini - Light & Fast)
google-generativeai>=0.8.3
Pillow>=10.0.0
requests>=2.31.0
# Utils
requests>=2.31.0
python-dateutil>=2.9.0.post0
python-docx>=1.1.0

Flask-Mail